Moved timing update routines from place.cpp to place_timing_update.*. Enhanced documentation.
