+++ libs/libvtrutil/src/vtr_vec_id_set.h
@@ -2,6 +2,7 @@
 #define VTR_SET_H
 
 #include <vector>
+#include <algorithm>
 
 namespace vtr {
